# doc/swagger/swagger.yaml
...
swagger: "2.0"
info:
  description: Gestionnaire de projets
  version: 1.0.0
  title: CDP
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
host: virtserver.swaggerhub.com
basePath: /edupland/CDP/1.0.0
...
schemes:
- https
paths:
  /user:
    get:
      tags:
      - User
      summary: Connexion d'un utilisateur
      description: Un utilisateur se connecte à l'outil
      parameters:
      - name: Login
        in: query
        description: Login de l'utilisateur pour se connecter
        required: true
        type: string
      - name: Password
        in: query
        description: Mot de passe de l'utilisateur pour se connecter
        required: true
        type: string
      responses:
        "200":
          description: Connexion réussie
        "401":
          description: Utilisateur ou mot de passe incorrect
        "500":
          description: Échec vérification identifiants
    post:
      tags:
      - User
      summary: Création d'un utilisateur
      description: Un nouvel utilisateur est créé
      consumes:
      - application/json
      produces:
      - application/json
      parameters:
      - name: Nom
        in: query
        description: Nom de famille de l'utilisateur
        required: true
        type: string
      - name: Prenom
        in: query
        description: Prénom de l'utilisateur
        required: true
        type: string
      - name: Mail
        in: query
        description: Adresse mail de l'utilisateur
        required: true
        type: string
      - name: Societe
        in: query
        description: Société à laquelle appartient l'utilisateur
        required: false
        type: string
      - name: Login
        in: query
        description: Login de l'utilisateur pour se connecter
        required: true
        type: string
      - name: Password
        in: query
        description: Mot de passe de l'utilisateur pour se connecter
        required: true
        type: string
      responses:
        "201":
          description: Inscription réussie
          schema:
            $ref: '#/definitions/User'
        "409":
          description: Login déjà existant
        "500":
          description: |-
            Échec vérification login
            Échec de l'inscription
            Connexion à la base de données impossible
            Fermeture base de donnée impossible
    put:
      tags:
      - User
      summary: Mettre à jour le profil
      description: Changer les informations d'un utilisateur
      parameters:
      - name: oldLogin
        in: query
        description: Login de l'utilisateur
        required: true
        type: string
      - name: Nom
        in: query
        description: Nouveau Nom
        required: false
        type: string
      - name: Prenom
        in: query
        description: Nouveau Prénom
        required: false
        type: string
      - name: Mail
        in: query
        description: Nouvelle adresse mail
        required: false
        type: string
      - name: Societe
        in: query
        description: Nouvelle société
        required: false
        type: string
      - name: Login
        in: query
        description: Nouveau login
        required: false
        type: string
      - name: Password
        in: query
        description: Nouveau mot de passe
        required: false
        type: string
      responses:
        "201":
          description: Mise à jour des informations de l'utilisateur réussie
          schema:
            $ref: '#/definitions/User'
        "409":
          description: Login déjà existant
        "500":
          description: |-
            Éche de la vérification de l'ancien login
            Échec de la mise à jour des informations de l'utilisateur
...
  /task:
    get:
      tags:
      - Task
      summary: Récupération des tâches
      description: On récupère les tâches d'un projet
      parameters:
      - name: IDProjet
        in: query
        description: Id du projet dont on veut les tâches
        required: true
        type: string
      responses:
        "200":
          description: La liste des tâches du projet dont l'id est IDProjet
        "500":
          description: |-
            Connexion à la base de données impossible
            Échec de la récupération des tâches
    post:
      tags:
      - Task
      summary: Création d'un tâche
      description: Une tâche est crée
      parameters:
      - name: IDProjet
        in: query
        description: Id du projet
        required: true
        type: string
      - name: Titre
        in: query
        description: Titre de la tâche
        required: true
        type: string
      - name: Statut
        in: query
        description: État de la tâche
        required: true
        type: string
        enum:
        - todo
        - pending
        - done
      responses:
        "201":
          description: Création de la tâche réussie
          schema:
            $ref: '#/definitions/Task'
        "500":
          description: |-
            Connexion à la base de données impossible
            Échec de la création de la tâche
...
definitions:
  User:
    type: object
    required:
    - Login
    - Mail
    - Nom
    - Password
    - Prenom
    properties:
      Nom:
        type: string
        example: Dupont
      Prenom:
        type: string
        example: Michel
      Mail:
        type: string
        example: michel.dupont@mail.fr
      Societe:
        type: string
      Login:
        type: string
        example: mdupont
      Password:
        type: string
        example: mdupont
...
  Task:
    type: object
    required:
    - Description
    - Projet
    - Statut
    - Titre
    properties:
      Projet:
        $ref: '#/definitions/Issue_Projet'
      Titre:
        type: string
        example: Titre de la tâche 1
      Description:
        type: string
        example: Voici la description de la tâche 1
      Statut:
        type: string
        example: todo
        enum:
        - todo
        - pending
        - done
...
  Project_Proprietaire:
    type: object
    required:
    - Login
    - Mail
    - Nom
    - Password
    - Prenom
    properties:
      Nom:
        type: string
        example: Dupont
      Prenom:
        type: string
        example: Michel
      Mail:
        type: string
        example: michel.dupont@mail.fr
      Societe:
        type: string
      Login:
        type: string
        example: mdupont
      Password:
        type: string
        example: mdupont
  Project_Developpeurs:
    type: object
    required:
    - Login
    - Mail
    - Nom
    - Password
    - Prenom
    properties:
      Nom:
        type: string
        example: Garcia
      Prenom:
        type: string
        example: Gabriel
      Mail:
        type: string
        example: gabriel.garcia@mail.fr
      Societe:
        type: string
        example: Capgemini
      Login:
        type: string
        example: ggarcia
      Password:
        type: string
        example: ggarcia
  Project_Clients:
    type: object
    required:
    - Login
    - Mail
    - Nom
    - Password
    - Prenom
    properties:
      Nom:
        type: string
        example: Martin
      Prenom:
        type: string
        example: Bernard
      Mail:
        type: string
        example: bernard.martin@mail.fr
      Societe:
        type: string
        example: Dassault
      Login:
        type: string
        example: bmartin
      Password:
        type: string
        example: bmartin
  Issue_Projet:
    type: object
    required:
    - Clients
    - Description
    - Developpeurs
    - Proprietaire
    - Titre
    properties:
      Titre:
        type: string
        example: Titre du projet
      Description:
        type: string
        example: Voici la description du projet
      Proprietaire:
        $ref: '#/definitions/Project_Proprietaire'
      Developpeurs:
        $ref: '#/definitions/Project_Developpeurs'
      Clients:
        $ref: '#/definitions/Project_Clients'